+++ kwimage/util_warp.py
@@ -85,11 +85,12 @@ def _coordinate_grid(dims, align_corners=False):
 
 
 def warp_image(inputs, mat, **kw):
-    import kwarray
+    raise AssertionError('do not use')
+    # import kwarray
     # _impl = kwarray.ArrayAPI.coerce(inputs)
-    inputs = kwarray.atleast_nd(inputs, 3)
-    tensor = inputs.transpose(2, 0, 1)
-    return warp_tensor(tensor, mat, **kw)
+    # inputs = kwarray.atleast_nd(inputs, 3)
+    # tensor = inputs.transpose(2, 0, 1)
+    # return warp_tensor(tensor, mat, **kw)
 
 
 def warp_tensor(inputs, mat, output_dims, mode='bilinear',